+++ database/migrations/2022_09_05_155406_create_payments_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->foreignId('added_by_id')->nullable()->index();
             $table->foreignId('approved_by_id')->nullable()->index();
             $table->string('status', 30); // enum
-            $table->float('value')->default(0);
+            $table->integer('value')->default(0);
             $table->text('note')->nullable();
             $table->timestamps();
             $table->softDeletes();
